requirements.txt
requests
python-telegram-bot
schedule
beautifulsoup4
instabot
aiohttp
Flask
Faker
user-agent
